Feature: Animate progress counters for day progress and centralize logic
This commit introduces an animated progress counter and refactors the way progress is calculated and displayed.

The `passages_completed` string has been updated to remove the numeric placeholders, allowing for a more flexible and animated presentation of the completion count (e.g., "1/5 passages completed").

A new generic `AnimatedIntText` composable has been created in the `ui/component` module. This component is now used in both the `WeekProgressLabel` and the `DayProgress` to animate changes in the progress count, providing a smoother user experience.

The logic for calculating passage and chapter completion, previously duplicated in the `DayProgress` composable, has been moved into the `DayUiStateFlowFactory`. This centralization simplifies the UI components, making them stateless and improving the overall architecture by separating concerns. The `DayUiState` now directly provides the completed and total passage counts to the UI.
