+++ feature/day/src/commonMain/kotlin/com/quare/bibleplanner/feature/day/presentation/component/DayScreenTopBarComponent.kt
@@ -83,8 +83,8 @@ internal fun DayScreenTopBarComponent(
                     }
                     VerticalSpacer(4)
                     DayProgress(
-                        passages = day.passages,
-                        books = books,
+                        completedCount = completedPassagesCount,
+                        totalCount = totalPassagesCount,
                     )
                 }
             }
