+++ feature/day/src/commonMain/kotlin/com/quare/bibleplanner/feature/day/presentation/factory/DayUiStateFlowFactory.kt
@@ -1,5 +1,8 @@
 package com.quare.bibleplanner.feature.day.presentation.factory
 
+import com.quare.bibleplanner.core.model.book.BookDataModel
+import com.quare.bibleplanner.core.model.plan.ChapterPlanModel
+import com.quare.bibleplanner.core.model.plan.PassagePlanModel
 import com.quare.bibleplanner.core.model.plan.ReadingPlanType
 import com.quare.bibleplanner.feature.day.domain.EditDaySelectableDates
 import com.quare.bibleplanner.feature.day.domain.usecase.CalculateAllChaptersReadStatusUseCase
@@ -72,6 +75,11 @@ internal class DayUiStateFlowFactory(
                 )
             }
 
+            val (completedCount, totalCount) = calculatePassageCounts(
+                passages = day.passages,
+                books = books,
+            )
+
             DayUiState.Loaded(
                 day = day,
                 weekNumber = weekNumber,
@@ -82,9 +90,85 @@ internal class DayUiStateFlowFactory(
                     passages = day.passages,
                     books = books,
                 ),
+                completedPassagesCount = completedCount,
+                totalPassagesCount = totalCount,
             )
         } else {
             DayUiState.Loading
+        }
+    }
+
+    /**
+     * Calculate the total number of chapters/items displayed and how many are completed.
+     * Returns a Pair of (completedCount, totalCount).
+     */
+    private fun calculatePassageCounts(
+        passages: List<PassagePlanModel>,
+        books: List<BookDataModel>,
+    ): Pair<Int, Int> {
+        var totalCount = 0
+        var completedCount = 0
+
+        passages.forEach { passage ->
+            if (passage.chapters.isEmpty()) {
+                // If no chapters specified, count as 1 item (the whole book)
+                totalCount++
+                if (passage.isRead) {
+                    completedCount++
+                }
+            } else {
+                // Count each chapter as a separate item
+                passage.chapters.forEach { chapter ->
+                    totalCount++
+                    val isChapterRead = isChapterReadForCount(
+                        passage = passage,
+                        chapter = chapter,
+                        books = books,
+                    )
+                    if (isChapterRead) {
+                        completedCount++
+                    }
+                }
+            }
+        }
+
+        return Pair(completedCount, totalCount)
+    }
+
+    /**
+     * Check if a specific chapter within a passage is read by checking the book data.
+     */
+    private fun isChapterReadForCount(
+        passage: PassagePlanModel,
+        chapter: ChapterPlanModel,
+        books: List<BookDataModel>,
+    ): Boolean {
+        val book = books.find { it.id == passage.bookId } ?: return false
+        val bookChapter = book.chapters.find { it.number == chapter.number } ?: return false
+
+        val startVerse = chapter.startVerse
+        val endVerse = chapter.endVerse
+
+        return when {
+            // If verse range is specified, check those specific verses
+            startVerse != null && endVerse != null -> {
+                val requiredVerses = startVerse..endVerse
+                requiredVerses.all { verseNumber ->
+                    bookChapter.verses.find { it.number == verseNumber }?.isRead == true
+                }
+            }
+
+            // If only start verse is specified, check from that verse to end of chapter
+            startVerse != null -> {
+                bookChapter.verses
+                    .filter { it.number >= startVerse }
+                    .all { it.isRead }
+            }
+
+            // If no verse range specified, check if entire chapter is read
+            else -> {
+                bookChapter.isRead
+            }
         }
     }
 }
