+++ feature/reading_plan/src/commonMain/kotlin/com/quare/bibleplanner/feature/readingplan/presentation/component/week/WeekProgressLabel.kt
@@ -1,13 +1,5 @@
 package com.quare.bibleplanner.feature.readingplan.presentation.component.week
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.ContentTransform
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.material3.MaterialTheme
@@ -18,9 +10,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import bibleplanner.feature.reading_plan.generated.resources.Res
 import bibleplanner.feature.reading_plan.generated.resources.week_progress_complete
+import com.quare.bibleplanner.ui.component.AnimatedIntText
 import org.jetbrains.compose.resources.stringResource
-
-private const val DAY_PROGRESS_ANIMATION_DURATION_MS = 300
 
 @Composable
 internal fun WeekProgressLabel(
@@ -31,19 +22,11 @@ internal fun WeekProgressLabel(
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.spacedBy(0.dp),
     ) {
-        AnimatedContent(
-            targetState = readDaysCount,
-            transitionSpec = {
-                getDayProgressAnimation()
-            },
+        AnimatedIntText(
+            value = readDaysCount,
             label = "readDaysCountAnimation",
-        ) { count ->
-            Text(
-                text = count.toString(),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Medium,
-            )
-        }
+            style = MaterialTheme.typography.titleMedium,
+        )
         Text(
             text = "/$totalDays",
             style = MaterialTheme.typography.titleMedium,
@@ -56,15 +39,3 @@ internal fun WeekProgressLabel(
         )
     }
 }
-
-private fun getDayProgressAnimation(): ContentTransform = (
-    fadeIn(animationSpec = tween(DAY_PROGRESS_ANIMATION_DURATION_MS)) + slideInVertically(
-        animationSpec = tween(DAY_PROGRESS_ANIMATION_DURATION_MS),
-        initialOffsetY = { it },
-    )
-) togetherWith (
-    fadeOut(animationSpec = tween(DAY_PROGRESS_ANIMATION_DURATION_MS)) + slideOutVertically(
-        animationSpec = tween(DAY_PROGRESS_ANIMATION_DURATION_MS),
-        targetOffsetY = { -it },
-    )
-)
